be/requirements.txt
Django==5.2.8
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
python-dotenv==1.0.0
python-socketio==5.10.0
uvicorn==0.30.0
psycopg2-binary==2.9.9
python-multipart==0.0.6
google-generativeai==0.3.0
